+++ CHANGELOG.md
@@ -27,7 +27,7 @@
     - Implement formatted logging
       - Allowing easier debugging
     - Implement per-version, per-run file logging
-      - ex. OpenCore-Patcher (0.6.7) (2021-12-31-12-34-56).log
+      - ex. OpenCore-Patcher (0.6.7) (2021-12-31-12-34-56-666903).log
     - Keep only 10 latest log files
     - Reveal log file in Finder on main thread crash
     - Avoid writing username to log file